requirements.txt
pytest>=7.0.0
pywinauto>=0.6.8
pyyaml>=6.0
openpyxl>=3.0.10
markdown>=3.4.1
beautifulsoup4>=4.11.1
allure-pytest>=2.13.2
Pillow>=9.3.0
